perf: int32 LUTs + software prefetch for stmap warp kernels
Three optimizations to the bilinear warp assembly kernels:

1. int32 LUTs instead of int64: halves LUT memory from 33MB to 16.6MB
   at 1080p. Values fit (max 1919*65536 = 126M < 2^31). Fewer cache
   misses when reading LUT entries.

2. Software prefetch (PRFM PLDL1KEEP on arm64, PREFETCHT0 on amd64):
   reads next pixel's sy LUT entry, computes its row address, and
   prefetches the source row into L1. This hides the dominant cost
   (random source pixel reads causing L2/L3 cache misses) by
   overlapping memory access with computation.

3. row1off = row0off + srcW: eliminates one MUL per pixel for the
   second row offset (iy1 = iy + 1, so row1 = row0 + stride).

Benchmark (1080p, M1 Pro):
  Before (int64, no prefetch): 14.7ms
  After  (int32 + prefetch):   10.3ms  (30% faster, zero allocs)

+++ server/stmap/warp_kernels_amd64.go
@@ -2,13 +2,14 @@
 
 package stmap
 
-// warpBilinearRow processes one band of pixels using precomputed 16.16
-// fixed-point LUT coordinates. For each output pixel i, reads lutX[i] and
-// lutY[i], samples 4 source pixels from src (full plane, srcW×srcH), and
-// writes the bilinear-interpolated result to dst[i].
+// warpBilinearRow processes n output pixels using precomputed 16.16
+// fixed-point LUT coordinates (int32). For each pixel, reads source
+// coordinates from lutX/lutY, samples 4 source pixels with bilinear
+// interpolation, and writes the result to dst.
 //
-// This eliminates Go bounds checks on the src array access and uses
-// register-pinned loop variables for ~2x speedup over the Go fallback.
+// Uses software prefetching to hide memory latency from the random
+// source pixel access pattern (each pixel reads from a different
+// source location, causing L2/L3 cache misses).
 //
 //go:noescape
-func warpBilinearRow(dst, src *byte, srcW, srcH, n int, lutX, lutY *int64)
+func warpBilinearRow(dst, src *byte, srcW, srcH, n int, lutX, lutY *int32)
